Extract ScoreTree classes to a trait.
